Refactor PCA visualization
